feat: Enable parsing of build slug from URL
Previously, the hull picker page served as a landing page for the app
and the build page could only be reached from there. With the
introduction of routes, the app needs to be able to render the correct
content based on the route, specifically the correct hull and build for the
build view.

- Adjusts the `ShipBuilder` to set the correct state, based on the URL.
- Makes sure child components are only rendered once the necessary
state has been set.

- Renames `Hull Select` to `HullPicker` and `SelectMenu` to
`HullSelect`, to be consistent with naming convention in `ShipBuilder`.